+++ src/pydrobert/kaldi/eval/__init__.py
@@ -14,4 +14,4 @@
 
 """Tools related to evaluating models"""
 
-__all__ = ["util"]
+__all__ = []